[enhance] Add flag for indication inlineScript is being used
This plugin prior to 1.5.0 did not perform any validation of input arguments before passing on required items to makeself and letting makeself fail if needed.  Some concern around that was raised specifically with our defaults and subsequently fixed as the entire plugin existence was designed to run with the 4 required parameters to makeself and not edge cases it supports.  Some use cases may include just directly running an inline script itself which can be very useful.  Therefore, added a flag specifically for that need.

Note: Makeself does allow for not strictly having the script at all in case of pure directory extraction which was likely added sometime after this plugin creation.  That seems useful but is not currently supported here.