[rework] Rename targetDirectory to makeselfTempDirectory
makeself has an extracting target directory with same name, so move this so its not so confusing.  The extraction directory will be named as extraction for clarity but need to avoid confusion as best possible.  This directory is only used for extraction of makeself scripts from the jar file during the build.  So simply calling it the directory its named and for which is ready only causes no impact to users.