Enhance makeself to log the plugin version used in creation - hard-coded!
most of this is not hard-coded but the MS_PLUGIN_VERSION is hard-coded, add release checklist to what to do before release which includes addressing this.  Because running the update will overwrite files this should be obvious but if that is skipped it will obviously be wrong but it will still be better than no indication at all as this seems helpful for later debugging purposes.  A better way could be written but this seemed quickest to get it in place.