Expose gitPath out to user for windows usage
Today on windows, this product uses portable git.  This was to overcome users originally not setting git up on the path which resulted in tickets for usage.  However, that comes with some overhead that is not generally necessary as most users likely are using git for windows on windows.  Therefore allow user to define where that is located so we can use the usr/bin from that location.   This still avoids the original setup issue, if not set, will simply defer to portable git.